+++ .gitignore
@@ -42,4 +42,4 @@ next-env.d.ts
 .cursorrules
 .vscode/
 
-public/sw.js
+sw.js
