+++ app/live/[address]/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-
+import { useState, useEffect } from "react";
 import { Broadcast, createStreamViaProxy } from "@/components/Live/Broadcast";
 import { useOrbisContext } from "@/context/OrbisContext";
 import {
@@ -17,7 +17,6 @@ import { MembershipGuard } from "@/components/auth/MembershipGuard";
 import { ProfilePageGuard } from "@/components/UserProfile/UserProfile";
 import { MultistreamTargetsForm } from "@/components/Live/multicast/MultistreamTargetsForm";
 import { MultistreamTargetsList } from "@/components/Live/multicast/MultistreamTargetList";
-import { useState, useEffect } from "react";
 import {
   listMultistreamTargets,
   MultistreamTarget,
