+++ app/profile/page.tsx
@@ -1,19 +1,23 @@
 "use client";
 import { useEffect } from "react";
-import { useUser } from "@account-kit/react";
+import { useUser, useSmartAccountClient } from "@account-kit/react";
 import { useRouter } from "next/navigation";
+import { Skeleton } from "@/components/ui/skeleton";
 
 function ProfileRedirect() {
   const router = useRouter();
+  const { address: scaAddress } = useSmartAccountClient({});
   const user = useUser();
+  const eoaAddress = user?.address;
 
   useEffect(() => {
-    if (user?.address) router.replace(`/profile/${user.address}`);
-  }, [user, router]);
+    if (eoaAddress) router.replace(`/profile/${eoaAddress || scaAddress}`);
+  }, [eoaAddress, scaAddress, router]);
 
   return (
-    <div className="flex items-center justify-center h-screen text-lg font-medium">
-      Loading profile...
+    <div className="flex flex-col items-center justify-center h-screen gap-4">
+      <Skeleton className="h-12 w-12 rounded-full" />
+      <Skeleton className="h-4 w-32 rounded" />
     </div>
   );
 }
