+++ app/vote/create/[address]/page.tsx
@@ -1,0 +1,49 @@
+import React from "react";
+import {
+  Breadcrumb,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbList,
+  BreadcrumbPage,
+  BreadcrumbSeparator,
+} from "@/components/ui/breadcrumb";
+import { Slash } from "lucide-react";
+import { Create } from "@/components/vote/Create";
+import { MembershipGuard } from "@/components/auth/MembershipGuard";
+
+const CreateVotePage = () => {
+  return (
+    <div className={"container"}>
+      <div className="my-5 p-4">
+        <Breadcrumb>
+          <BreadcrumbList>
+            <BreadcrumbItem>
+              <BreadcrumbLink href="/">
+                <span role="img" aria-label="home">
+                  🏠
+                </span>{" "}
+                Home
+              </BreadcrumbLink>
+            </BreadcrumbItem>
+            <BreadcrumbSeparator>
+              <Slash />
+            </BreadcrumbSeparator>
+            <BreadcrumbItem>
+              <BreadcrumbLink href="/vote">Vote</BreadcrumbLink>
+            </BreadcrumbItem>
+            <BreadcrumbSeparator>
+              <Slash />
+            </BreadcrumbSeparator>
+            <BreadcrumbItem>
+              <BreadcrumbPage>Create</BreadcrumbPage>
+            </BreadcrumbItem>
+          </BreadcrumbList>
+        </Breadcrumb>
+      </div>
+      <MembershipGuard>
+        <Create />
+      </MembershipGuard>
+    </div>
+  );
+};
+export default CreateVotePage;
